moved cli to a public location
